requirement.txt
streamlit
time
ollama
pillow